added the 'createComment' method to WallService
